requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
python-dotenv==1.2.1
    # via -r requirements.in
